+++ src/iec-resolver.ts
@@ -24,52 +24,64 @@ SOFTWARE.
 
 
 
-import * as types from "./iec-types"
+import * as iecTypes from "./iec-types"
 
-import { ExtractedStruct, ExtractedStructVariable, IecType } from "./types/types"
+import {
+  dataTypeUnit,
+  EnumEntry,
+  ExtractedEnum,
+  ExtractedType,
+  IecType,
+  ExtractedAlias,
+  EnumDataType,
+  ExtractedTypeVariable
+} from "./types/types"
 
 
 /**
  * Available non-complex IEC types
  */
 const nonComplexTypes = [
-  types.BOOL,
-  types.USINT,
-  types.BYTE,
-  types.SINT,
-  types.UINT,
-  types.WORD,
-  types.INT,
-  types.DINT,
-  types.UDINT,
-  types.DWORD,
-  types.TIME,
-  types.TOD,
-  types.TIME_OF_DAY,
-  types.DT,
-  types.DATE_AND_TIME,
-  types.DATE,
-  types.REAL,
-  types.LREAL,
-  types.ULINT,
-  types.LWORD,
-  types.LINT
+  iecTypes.BOOL,
+  iecTypes.USINT,
+  iecTypes.BYTE,
+  iecTypes.SINT,
+  iecTypes.UINT,
+  iecTypes.WORD,
+  iecTypes.INT,
+  iecTypes.DINT,
+  iecTypes.UDINT,
+  iecTypes.DWORD,
+  iecTypes.TIME,
+  iecTypes.TOD,
+  iecTypes.TIME_OF_DAY,
+  iecTypes.DT,
+  iecTypes.DATE_AND_TIME,
+  iecTypes.DATE,
+  iecTypes.REAL,
+  iecTypes.LREAL,
+  iecTypes.ULINT,
+  iecTypes.LWORD,
+  iecTypes.LINT
 ]
 
-
-
 /**
- * RegExp pattern for matching TYPEs like STRUCT etc.
- *  TYPE
- *   ...
- *  END_TYPE
+ * RegExp pattern for matching data type units DUTs (struct, union, alias, enum)
  */
-const typeRegEx = new RegExp(/type\s*(\w*)\s*:(.*?)end_type/gis)
+const typeRegEx = new RegExp(/type\s*(\w*)\s*[:]*\s*(struct|union|\(|:)\s*(.*?)(?:end_struct|end_union|;|\)\s*([^\s]*?)\s*;)\s*end_type/gis)
+
+
+
 
 /**
  * RegExp pattern for matching STRUCT variables (children)
  */
 const structVariableRegEx = new RegExp(/(\w+)\s*:\s*([^:;]*)/gis)
+
+/**
+ * RegExp pattern for matching ENUM
+ */
+const enumRegEx = new RegExp(/\s*(.+?)\s*(?::=\s*(.*?)\s*)*(?:,|$)/gis)
 
 
 /**
@@ -97,47 +109,88 @@ const arrayRegEx = new RegExp(/array\s*[\[(]+(.*?)[\])]\s*of\s*([^:;]*)/gis)
 const arrayDimensionsRegEx = new RegExp(/(?:\s*(?:([^\.,\s]*)\s*\.\.\s*([^,\s]*))\s*)/gis)
 
 
+
+
+
 /**
  * Extracts struct declarations from given string containing one or multiple TYPE...END_TYPE declarations
- * @param declaration 
+ * @param declaration Declaration string
+ * @returns Extracted data types
  */
-const extractStructDeclarations = (declarations: string): ExtractedStruct[] => {
-
-  const extractedStructs: ExtractedStruct[] = []
+const extractTypeDeclarations = (declarations: string): ExtractedType[] => {
+  const extractedTypes: ExtractedType[] = []
   let match: RegExpExecArray | null
   const typeMatcher = new RegExp(typeRegEx)
-  
+
   //Looping until all no more declarations found
   //TODO: Add checks if RegExp was successful
   while ((match = typeMatcher.exec(declarations)) !== null) {
-
     // This is necessary to avoid infinite loops with zero-width matches
     if (match.index === typeMatcher.lastIndex) {
       typeMatcher.lastIndex++;
     }
 
-    //Creating new temporary extracted struct object
-    extractedStructs.push({
-      dataType: match[1],
-      children: extractStructVariables(match[2]),
+    if (match.length < 5) {
+      throw new Error(`Problem extracting IEC type declaration from given string. RegExp result has less than 4 matches: ${JSON.stringify(match)}`)
+    }
+
+    const type = {
       resolved: undefined
-    })
+    } as ExtractedType
+
+    //Match 1 is the user-defined name
+    type.name = match[1]
+
+    //Match 2 provides info which type is it
+    //Match 3 is the content (depends on type)
+    switch (match[2].toLowerCase()) {
+      //STRUCT:
+      case 'struct':
+        type.type = dataTypeUnit.STRUCT
+        type.content = extractTypeVariables(match[3])
+        break;
+
+      //UNION:
+      case 'union':
+        type.type = dataTypeUnit.UNION
+        type.content = extractTypeVariables(match[3])
+        break;
+
+      //ENUM:
+      case '(':
+        type.type = dataTypeUnit.ENUM
+        type.content = extractEnum(match[3], match[4])
+        break;
+      
+      //ALIAS:
+      case ':':
+        type.type = dataTypeUnit.ALIAS
+        type.content = {
+          dataType: match[3]
+        } as ExtractedAlias
+        break;
+      
+      default:
+        throw new Error(`Problem extracting IEC data type (DUT) from given string. Found match: ${JSON.stringify(match)}`)
+    }
+
+    extractedTypes.push(type)
   }
 
-
-  return extractedStructs
+  return extractedTypes
 }
 
 
 
+
 /**
- * Extracts struct variables (children) from given declaration string
+ * Extracts STRUCT/UNION variables (children) from given declaration string
  * @param declaration 
  * @returns 
  */
-const extractStructVariables = (declaration: string): ExtractedStructVariable[] => {
+const extractTypeVariables = (declaration: string): ExtractedTypeVariable[] => {
 
-  const extractedVariables: ExtractedStructVariable[] = []
+  const extractedVariables: ExtractedTypeVariable[] = []
 
   let match: RegExpExecArray | null
   const structVariableMatcher = new RegExp(structVariableRegEx)
@@ -162,77 +215,168 @@ const extractStructVariables = (declaration: string): ExtractedStructVariable[] 
 
 
 
+/**
+ * Extracts ENUM from given declaration string
+ * @param declaration 
+ * @returns 
+ */
+const extractEnum = (declaration: string, dataType?: string): ExtractedEnum => {
+
+  if (dataType === undefined || dataType === '')
+    dataType = 'INT'
+  
+  const extractedEnum = {
+    dataType,
+    content: {}
+  } as ExtractedEnum
+
+  let match: RegExpExecArray | null
+  const enumMatcher = new RegExp(enumRegEx)
+  const enumList = [] as EnumEntry[]
+
+  //TODO: Add checks if RegExp was successful
+  while ((match = enumMatcher.exec(declaration)) !== null) {
+
+    // This is necessary to avoid infinite loops with zero-width matches
+    if (match.index === enumMatcher.lastIndex) {
+      enumMatcher.lastIndex++
+    }
+
+    let value = match[2] === undefined ? undefined : parseInt(match[2])
+    
+    //If value is undefined, it's not provided in the ENUM
+    //--> The value is previous value + 1 if available
+    //Otherwise value is 0
+    if (value === undefined && enumList.length > 0) {
+      value = enumList[enumList.length - 1].value + 1
+      
+    } else if(value === undefined){
+      value = 0
+    }
+
+    if (value === undefined)
+      throw new Error(`Problem calculating ENUM entry "${match[1]}". Found match: ${JSON.stringify(match)}`)
+
+    //TODO: When adding support for constants edit this
+    //We could have ENUM that has value := SOME_CONSTANT
+    if (isNaN(value))
+      throw new Error(`Problem calculating ENUM entry "${match[1]}" value from "${match[2]}". Found match: ${JSON.stringify(match)}`)
+
+    enumList.push({
+      name: match[1],
+      value
+    })
+
+    extractedEnum.content[match[1]] = value
+  }
+
+  return extractedEnum
+}
+
+
+
 
 /**
- * Resolves given string PLC STRUCT declaration to IEC data types
+ * Resolves given string PLC type declarations to IEC data types
  * @param declarations 
  * @param topLevelDataType 
  * @param providedTypes 
  * @returns 
  */
-export const resolveIecStructs = (declarations: string, topLevelDataType?: string, providedTypes?: Record<string, IecType>): IecType => {
-  //First extracting struct definitions from string
-  const structs = extractStructDeclarations(declarations)
+export const resolveIecTypes = (declarations: string, topLevelDataType?: string, providedTypes?: Record<string, IecType>): IecType => {
+  //First extracting basic type definitions from string
+  const types = extractTypeDeclarations(declarations)
 
-  //If multiple structs found, we need to know which one is the top-level
-  if (!topLevelDataType && structs.length > 1) {
-    throw new Error('Top level data type name (topLevelDataType) is not given and many struct declarations found. Not possible to guess.')
+  //If multiple data types found, we need to know which one is the top-level
+  if (!topLevelDataType && types.length > 1) {
+    throw new Error('Top level data type name (topLevelDataType) is not given and multiple type declarations found. Not possible to guess.')
 
   } else if (!topLevelDataType) {
-    //When only one struct type found, we know it's the top-level
-    topLevelDataType = structs[0].dataType
+    //When only one type found, we know it's the top-level
+    topLevelDataType = types[0].name
   }
-
-  //Resolving structs to IEC data types
-  for (const struct of structs) {
-    //If already resolved, skip (happens if some other struct already depended on it)
-    if (struct.resolved)
+  
+  //Resolving types to IEC data types
+  for (const type of types) {
+    //If already resolved, skip (happens if some other type already depended on it)
+    if (type.resolved)
       continue
     
-    struct.resolved = resolveIecStruct(struct, structs, providedTypes)
+    type.resolved = resolveIecDataTypeUnit(type, types, providedTypes)
   }
 
   //Return the top-level struct
-  const returnVal = structs.find(struct => topLevelDataType !== undefined && struct.dataType.toLowerCase() === topLevelDataType.toLowerCase())
+  const returnVal = types.find(type => topLevelDataType !== undefined && type.name.toLowerCase() === topLevelDataType.toLowerCase())
 
   if (!returnVal) {
-    throw new Error(`Top-level type ${topLevelDataType} was not found - Is data type declaration provided? Types found: ${structs.map(struct => struct.dataType).join(', ')}`)
+    throw new Error(`Top-level type ${topLevelDataType} was not found - Is data type declaration provided? Types found: ${types.map(type => type.name).join(', ')}`)
   }
 
-  return types.STRUCT(returnVal.resolved)
+  return returnVal.resolved as IecType
 }
 
 
+ 
 
 
 /**
- * Resolves a single struct to IEC data type 
- * @param struct 
- * @param structs 
- * @param providedTypes 
- * @returns 
+ * Resolves single IEC data type unit from given parsed ExtractedType object
+ * 
+ * @param type Type to be resolved
+ * @param types List of all available types that are/will be resolved
+ * @param providedTypes List of user-provided types
+ * @returns Resolved type (IecType object)
  */
-const resolveIecStruct = (struct: ExtractedStruct, structs: ExtractedStruct[], providedTypes?: Record<string, IecType>): Record<string, IecType> => {
-  const resolved = {} as Record<string, IecType>
+const resolveIecDataTypeUnit = (type: ExtractedType, types: ExtractedType[], providedTypes?: Record<string, IecType>): IecType => {
 
-  for (const variable of struct.children) {
-    resolved[variable.name] = resolveIecVariable(variable.dataType, structs, providedTypes)
+  switch (type.type) {
+    case dataTypeUnit.STRUCT: {
+      const children = {} as Record<string, IecType>
+
+      for (const variable of type.content as ExtractedTypeVariable[]) {
+        children[variable.name] = resolveIecVariable(variable.dataType, types, providedTypes)
+      }
+      return iecTypes.STRUCT(children)
+    }
+
+    case dataTypeUnit.UNION: {
+      //Basically 1:1 as struct
+      const children = {} as Record<string, IecType>
+
+      for (const variable of type.content as ExtractedTypeVariable[]) {
+        children[variable.name] = resolveIecVariable(variable.dataType, types, providedTypes)
+      }
+      return iecTypes.UNION(children)
+    }
+    
+    case dataTypeUnit.ENUM: {
+      const enumeration = type.content as ExtractedEnum
+      
+      //TODO: Check if valid ENUM data type
+      return iecTypes.ENUM(enumeration.content, resolveIecVariable(enumeration.dataType, types, providedTypes) as EnumDataType)
+    }
+
+    case dataTypeUnit.ALIAS:
+      return resolveIecVariable((type.content as ExtractedAlias).dataType, types, providedTypes)
+
+    default:
+      throw new Error(`Problem resolving IEC data type unit (DUT). Unknown type: ${type.type}`)
   }
-
-  return resolved
 }
 
 
 
+
+
 /**
- * Resolves a single variable to IEC data type
+ * Resolves a single variable data type (string) to IEC data type
  * Calls itself recursively if needed (like array types)
- * @param dataType 
- * @param structs 
- * @param providedTypes 
- * @returns 
+ * @param dataType Data type name as string
+ * @param structs List of all available types that are/will be resolved
+ * @param providedTypes List of user-provided types
+ * @returns Resolved type (IecType object)
  */
-const resolveIecVariable = (dataType: string, structs: ExtractedStruct[], providedTypes?: Record<string, IecType>): IecType => {
+const resolveIecVariable = (dataType: string, types: ExtractedType[], providedTypes?: Record<string, IecType>): IecType => {
   //Simple non-complex data type
   let type: IecType | undefined = nonComplexTypes.find(type => type.type.toLowerCase() === dataType.toLowerCase())
 
@@ -247,10 +391,10 @@ const resolveIecVariable = (dataType: string, structs: ExtractedStruct[], provid
 
   if (stringMatch) {
     if (stringMatch[1].toLowerCase() === 'string') {
-      return types.STRING(stringMatch[3] ? parseInt(stringMatch[3]) : 80)
+      return iecTypes.STRING(stringMatch[3] ? parseInt(stringMatch[3]) : 80)
 
     } else if (stringMatch[1].toLowerCase() === 'wstring') {
-      return types.WSTRING(stringMatch[3] ? parseInt(stringMatch[3]) : 80)
+      return iecTypes.WSTRING(stringMatch[3] ? parseInt(stringMatch[3]) : 80)
 
     } else {
       throw new Error(`Unknown STRING definition: "${stringMatch}"`)
@@ -263,7 +407,7 @@ const resolveIecVariable = (dataType: string, structs: ExtractedStruct[], provid
   const arrayMatch = arrayMatcher.exec(dataType)
 
   if (arrayMatch) {
-    type = resolveIecVariable(arrayMatch[2], structs, providedTypes)
+    type = resolveIecVariable(arrayMatch[2], types, providedTypes)
   
     if (!type) {
       //This shouldn't happen
@@ -285,33 +429,33 @@ const resolveIecVariable = (dataType: string, structs: ExtractedStruct[], provid
 
       dimensions.push(parseInt(match[2]) - parseInt(match[1]) + 1)
     }
-    return types.ARRAY(type, dimensions)
+    return iecTypes.ARRAY(type, dimensions)
   }
 
 
-  //Struct (or unknown)
-  const struct = structs.find(struct => struct.dataType.toLowerCase() == dataType.toLowerCase())
+  //Data type unit (DUT) like struct, enum etc. (or unknown)
+  const dataTypeUnit = types.find(type => type.name.toLowerCase() == dataType.toLowerCase())
 
-  if (struct) {
-    //If struct is found but not yet resolved, resolve it now
-    if (!struct.resolved) {
-      struct.resolved = resolveIecStruct(struct, structs, providedTypes)
+  if (dataTypeUnit) {
+    //If type is found but not yet resolved, resolve it now
+    if (!dataTypeUnit.resolved) {
+      dataTypeUnit.resolved = resolveIecDataTypeUnit(dataTypeUnit, types, providedTypes)
     }
 
-    return types.STRUCT(struct.resolved)
+    return dataTypeUnit.resolved as IecType
 
   } else {
-    //Struct type was unknown. Was it provided already?
+    //Data type was unknown. Was it provided already?
     if (providedTypes) {
       const key = Object.keys(providedTypes).find(key => key.toLowerCase() === dataType.toLowerCase())
 
       if (key) {
         return providedTypes[key]
       } else {
-        throw new Error(`Unknown data type "${dataType}" found! Types found from declaration: ${structs.map(struct => struct.dataType).join(', ')}, types provided separately: ${Object.keys(providedTypes).join(',')}`)
+        throw new Error(`Unknown data type "${dataType}" found! Types found from declaration: ${types.map(type => type.name).join(', ')}, types provided separately: ${Object.keys(providedTypes).join(',')}`)
       }
     } else {
-      throw new Error(`Unknown data type "${dataType}" found! Types found from declaration: ${structs.map(struct => struct.dataType).join(', ')}`)
+      throw new Error(`Unknown data type "${dataType}" found! Types found from declaration: ${types.map(type => type.name).join(', ')}`)
     }
   }
 }