+++ src/iec-type-handler.ts
@@ -25,8 +25,10 @@ SOFTWARE.
 import iconv from 'iconv-lite'
 
 import type {
-  IecType,
-  StructChildren
+  EnumDataType,
+  EnumEntry,
+  EnumValue,
+  IecType
 } from './types/types'
 
 
@@ -50,21 +52,25 @@ export class STRUCT extends TypeBase implements IecType {
   /**
    * STRUCT children data types
    */
-  children: StructChildren
+  children?: Record<string, IecType | never>
 
-  constructor(children: StructChildren) {
+  constructor(children?: Record<string, IecType | never>) {
     super()
     
     this.children = children
 
     //Calculating struct size
     for (const key in this.children) {
+      if (typeof this.children[key] !== 'object' || this.children[key].byteLength === undefined) {
+        throw new Error(`Struct member ${key} is not valid IEC data type - Did you remember to use () with some data types that require it (example with STRING())?`)
+      }
+
       this.byteLength += this.children[key].byteLength
     }
   }
 
 
-  convertToBuffer(data: StructChildren): Buffer {
+  convertToBuffer(data: Record<string, unknown>): Buffer {
     if (!data)
       return Buffer.alloc(0)
 
@@ -88,6 +94,77 @@ export class STRUCT extends TypeBase implements IecType {
     for (const key in this.children) {
       obj[key] = this.children[key].convertFromBuffer(data.slice(pos, pos + this.children[key].byteLength))
       pos += this.children[key].byteLength
+    }
+
+    return obj
+  }
+
+  getDefault(): Record<string, unknown> {
+    const obj = {} as Record<string, unknown>
+
+    for (const key in this.children) {
+      obj[key] = this.children[key].getDefault()
+    }
+
+    return obj
+  }
+}
+
+
+
+/**
+ * IEC 61131-3 type: UNION
+ */
+export class UNION extends TypeBase implements IecType {
+  type = 'UNION'
+  /**
+   * UNION children data types
+   */
+  children?: Record<string, IecType | never>
+
+  constructor(children?: Record<string, IecType | never>) {
+    super()
+
+    this.children = children
+
+    //Calculating union size (= biggest child)
+    for (const key in this.children) {
+      if (typeof this.children[key] !== 'object' || this.children[key].byteLength === undefined) {
+        throw new Error(`Struct member ${key} is not valid IEC data type - Did you remember to use () with some data types that require it (example with STRING())?`)
+      }
+
+      if (this.children[key].byteLength > this.byteLength)
+        this.byteLength = this.children[key].byteLength
+    }
+  }
+
+
+  convertToBuffer(data: Record<string, unknown>): Buffer {
+    if (!data)
+      return Buffer.alloc(0)
+
+    //As UNION type member all are located in same memory, it's not that easy
+    //For now: Use the last given object key value
+    
+    const buffer = Buffer.alloc(this.byteLength)
+
+    for (const key in this.children) {
+      if (data[key] === undefined)
+        continue
+      
+      //There is only one value allowed so quit after this
+      const converted = this.children[key].convertToBuffer(data[key])
+      converted.copy(buffer, 0)
+    }
+
+    return buffer
+  }
+
+  convertFromBuffer(data: Buffer): Record<string, unknown> {
+    const obj = {} as Record<string, unknown>
+
+    for (const key in this.children) {
+      obj[key] = this.children[key].convertFromBuffer(data.slice(0, this.children[key].byteLength))
     }
 
     return obj
@@ -221,6 +298,127 @@ export class ARRAY extends TypeBase implements IecType {
   }
 }
 
+
+
+/**
+ * IEC 61131-3 type: ENUM
+ */
+export class ENUM extends TypeBase implements IecType {
+  type = 'ENUM'
+
+  definition: Record<string, number>
+  dataType: IecType
+
+  constructor(definition: Record<string, number>, dataType?: EnumDataType) {
+    super()
+
+    this.definition = definition
+
+    this.dataType = dataType ? dataType : new INT()
+    this.byteLength = this.dataType.byteLength
+  }
+
+
+  convertToBuffer(data: EnumValue): Buffer {
+    if (!data)
+      return Buffer.alloc(0)
+
+    if (typeof data === 'string') {
+      //Enumeration name given as string
+      const found = Object.keys(this.definition).find(key => key.toLowerCase() === data.toLowerCase())
+
+      if (found) {
+        return this.dataType.convertToBuffer(this.definition[found])
+      }
+      throw new Error('Input parameter is not valid for ENUM')
+
+    } else if (typeof data === 'number') {
+      //Enumeration value given as number
+      return this.dataType.convertToBuffer(data)
+
+    } else if (typeof data === 'object' && (data as EnumEntry).value) {
+      //Object given with value key
+      return this.dataType.convertToBuffer(data)
+
+    } else if (typeof data === 'object' && (data as EnumEntry).name) {
+      //Object given with name key
+      const found = Object.keys(this.definition).find(key => data.name && key.toLowerCase() === data.name.toLowerCase())
+
+      if (found) {
+        return this.dataType.convertToBuffer(this.definition[found])
+      }
+      throw new Error('Input parameter is not valid for ENUM')
+
+    } else {
+      throw new Error('Input parameter is not valid for ENUM (number, string or object with { value })')
+    }
+  }
+  
+
+  convertFromBuffer(data: Buffer): EnumEntry {
+    //First, converting buffer to number
+    const value = this.dataType.convertFromBuffer(data) as number
+
+    const entry = this.findEnumEntryByValue(value)
+
+    if (entry)
+      return entry
+
+    //Not found
+    return {
+      name: undefined,
+      value
+    }
+  }
+
+
+
+  getDefault(): EnumEntry {
+    //Codeys initializes the value with the first enumeration component
+    //Use it, unless there are none
+    const keys = Object.keys(this.definition)
+
+    if (keys.length > 0) {
+      return {
+        name: keys[0],
+        value: this.definition[keys[0]]
+      }
+
+    } else {
+      //No entries? Use data type default value
+      const value = (this.dataType as EnumDataType).getDefault()
+
+      //Do we have enumeration entry for default value?
+      const entry = this.findEnumEntryByValue(value)
+    
+      if (entry)
+        return entry
+    
+      //Not found
+      return {
+        name: undefined,
+        value
+      }
+    }
+  }
+
+
+
+  findEnumEntryByValue(value: number): EnumEntry | undefined {
+    for (const key in this.definition) {
+      if (this.definition[key] === value) {
+        //Found
+        return {
+          name: key,
+          value: value
+        }
+      }
+    }
+
+    //Not found
+    return undefined
+  }
+}