+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 17.08.2021
+### Added
+- `fromString()` support for other data types than STRUCT
+  - ENUM
+  - ALIAS
+  - UNION
+- New data types
+  - ENUM
+  - UNION
+  - ALIAS (only with `fromString()`)
+  
+### Updated
+- README updated
+- `fromString()` updated to work better
+
 ## [0.2.0] - 10.08.2021
 ### Added
 - fromString() method to convert IEC types automatically from PLC data type declarations